Added search by name API
Added all strains list API
Save strains in DB with descriptions

+++ server/src/main/java/ehealth/data_objects/UsageHistory.java
@@ -6,8 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
-import javax.persistence.Column;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -30,10 +28,10 @@ public class UsageHistory {
     private Integer strainId;
 
     @JsonProperty("medical_rank")
-    private Integer medicalRank;
+    private Double medicalRank;
 
     @JsonProperty("positive_rank")
-    private Integer positiveRank;
+    private Double positiveRank;
 
     @JsonProperty("overall_rank")
     private Integer overallRank;
